Add status endpoint
This enpoint provides the global game status.  It returns the following
information in a JSON formatted blob:

currentplayer => The name of the current player, as a string.
currentavg => The latest computed average score, as a string.
running => A string that says true or false, which indicates whether we
are current running the interactive component of the piece.