Unfuck random generator (255->256)
This was a bug that would have made whole regions of the RNG space
unevenly distributed and would have made it much more likely for those
regions to be collided.  Which, I mean, we don't have a secure model
here, we just store this shit in an insecure cookie and read back
whatever they've got.  But let's fix it anyway.