Add additional validation in rate.cgi
Prevent users from setting ratings over 5.  This makes it slightly
harder to manipulate the averages, in that you actually have to submit
multiple ratings to do it, instead of just submitting a bogus rating of
a particularly high value.  This resolves the bug documented in the
commit message for the introduction of the rate endpoint.