Update score (fixes #19)
This pushes a script that can be run on the server or from cron that
updates the average score every five seconds.  For some reason, the
luasql driver absolutely does not want to do multiple SQL statements in
an exec and I couldn't quite figure out how transactions were ever
supposed to work either.  (Dear luasql author, more than one page of
documentation on your entire SQL library would have been dearly
appreciated.)

Anyway, the resulting situation is that sometimes the world is just
unsafe.  And sometimes there's not going to be a currentavg in the
status API.  And so we really come down to the reality of the situation,
and the fundamental truths that we all seek to expose in art:

Life is unsafe.
Safety is an illusion.
Race conditions are real.
Try it from the ground first.
If you cry, you have to eat.